% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_pathways.R
\name{concat_pathways}
\alias{concat_pathways}
\title{Concatenate Pathway analysis report tables into one}
\usage{
concat_pathways(pathway_dir = NULL,
  pathway_pattern = "mcg_pathwayanalysis_pathway.xlsx",
  ouput_dir = NULL, ...)
}
\arguments{
\item{pathway_dir}{directory for pathways}

\item{pathway_pattern}{pathway pattern}

\item{ouput_dir}{ouput directory}
}
\description{
\code{concat_pathways} saved the result into a csv file
}
\author{
Samantha Shen \email{<shen.samantha@yahoo.com>}

License: GPL (>= 3)
}
