% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pathway_ids.R
\name{match_pathway_ids}
\alias{match_pathway_ids}
\title{Concatenate Pathway analysis report tables into one}
\usage{
match_pathway_ids(pathway_res, pathway_EID, pathway_raw)
}
\arguments{
\item{pathway_res}{pathway_res}

\item{pathway_EID}{pathway_EID}

\item{pathway_raw}{pathway_raw}
}
\description{
\code{concat_pathways} saved the result into a csv file
}
\author{
Samantha Shen \email{<shen.samantha@yahoo.com>}

License: GPL (>= 3)
}
